Выполняет неуспешную попытку всё исправить
