+++ package.json
@@ -19,6 +19,7 @@
     "eslint-config-htmlacademy": "2.1.0",
     "eslint-plugin-react": "7.22.0",
     "file-loader": "6.2.0",
+    "react-app-rewired": "^2.1.8",
     "style-loader": "2.0.0",
     "webpack": "4.46.0",
     "webpack-cli": "3.3.12",
