+++ src/components/offer/offer-gallery.jsx
