+++ src/mocks/offers.js
@@ -7,7 +7,7 @@ export default [
         "longitude": 4.895168,
         "zoom": 10,
       },
-      "name": `Amsterdam`,
+      "name": `Kyiv`,
     },
     "description": [
       `A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.`,
