+++ src/store/reducer.js
@@ -1,23 +1,11 @@
 import {ActionType} from "./action";
-import {LOCATIONS} from "../const";
-import {getOffersByCity} from "../common";
 
-const initialState = {
-  city: LOCATIONS[0].name,
-  rentalOffers: getOffersByCity(LOCATIONS[0].name),
-};
-
-const reducer = (state = initialState, action) => {
+const reducer = (state, action) => {
   switch (action.type) {
-    case ActionType.UPDATE_CITY:
+    case ActionType.SET_CITY:
       return {
         ...state,
         city: action.payload,
-      };
-    case ActionType.UPDATE_OFFERS:
-      return {
-        ...state,
-        rentalOffers: action.payload,
       };
     default:
       return state;
