+++ webpack.config.js
@@ -32,6 +32,9 @@ module.exports = {
     ],
   },
   resolve: {
+    alias: {
+      src: path.join(__dirname, 'src'),
+    },
     extensions: [".js", ".jsx"],
   },
   devtool: "source-map",
